fix: layout and components based on feedback
