+++ common/constants/partners.js
@@ -1,0 +1,44 @@
+const partners = [
+  {
+    name: 'APEX Systems',
+    logo: 'partnerLogos/apex_systems.png',
+    url: 'https://apexsystems.com',
+  },
+  {
+    name: 'GitHub',
+    logo: 'partnerLogos/github.png',
+    url: 'https://github.com',
+  },
+  {
+    name: '1Password',
+    logo: 'partnerLogos/1password.png',
+    url: 'https://1password.com/',
+  },
+  {
+    name: 'Docker',
+    logo: 'partnerLogos/docker.png',
+    url: 'https://docker.com',
+  },
+  {
+    name: 'Oracle',
+    logo: 'partnerLogos/oracle.png',
+    url: 'https://oracle.com',
+  },
+  {
+    name: "O'Reilly Media",
+    logo: 'partnerLogos/oreilly.png',
+    url: 'https://oreilly.com',
+  },
+  {
+    name: 'HackerRank',
+    logo: 'partnerLogos/hackerrank.png',
+    url: 'https://hackerrank.com',
+  },
+  {
+    name: 'Zapier',
+    logo: 'partnerLogos/zapier.png',
+    url: 'https://zapier.com',
+  },
+];
+
+export default partners;
