--- components/Partners/Partners.css
@@ -1,7 +1,0 @@
-.partnerLogos {
-  display: flex;
-  align-self: center;
-  flex-wrap: wrap;
-  justify-content: space-evenly;
-  margin: auto;
-}
